+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 - Added support for operator chaining (e.g. `a > b > c`).
 - Added support for `shortestPath` and `allShortestPaths` pattern constructs.
+- Added support for shortest path constructs (`SHORTEST k`, `ALL SHORTEST`, `SHORTEST k GROUPS`, and `ANY`).
 
 ### Changed